Improve the react asset integration
