+++ web/assets/js/react/index.js
@@ -1,19 +1,7 @@
 import { define } from "remount";
-import AppCheckClaim from "./app-check-claim";
-import AppSendClaim from "./app-send-claim";
+import AppAgreementSubmit from "./app-agreement-submit";
 
 define(
-    { "x-app-check-claim": AppCheckClaim },
+    { "x-app-agreement-submit": AppAgreementSubmit },
     { attributes: ["network"] }
-);
-
-define(
-    { "x-app-send-claim": AppSendClaim },
-    {
-        attributes: [
-            "network",
-            "proxy_contract_address",
-            "merkle_proof",
-            "amount"]
-    }
 );
